+++ core/src/test/java/com/jscisco/lom/domain/attribute/AttributeFactory.java
@@ -6,11 +6,11 @@ import com.jscisco.lom.domain.Name;
 public class AttributeFactory {
 
     public static Attribute testAttribute() {
-        return testAttribute(10);
+        return testAttribute(10f);
     }
 
 
-    public static Attribute testAttribute(int value) {
+    public static Attribute testAttribute(float value) {
         return new Attribute(Name.of("Test"), Description.of("Test"), value);
     }
